+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Fix "Nothing to do" always being printed.
+- Add experimental partial-staging support for GJF-managed Java files behind `KEMPT_EXPERIMENTAL_PARTIAL_GJF`.
 
 ## [0.0.4]